Refactor CSV exporter to separate file
