+++ src/features/grading/GradingUtils.ts
@@ -1,0 +1,62 @@
+import { GradingOverview } from './GradingTypes';
+
+export const exportGradingCSV = (gradingOverviews: GradingOverview[] | undefined) => {
+  if (!gradingOverviews) return;
+
+  const win = document.defaultView || window;
+  if (!win) {
+    console.warn('There is no `window` associated with the current `document`');
+    return;
+  }
+
+  const content = new Blob(
+    [
+      '"Assessment Number","Assessment Name","Student Name","Student Username","Group","Status","Grading","Question Count","Questions Graded","Initial XP","XP Adjustment","Current XP (excl. bonus)","Max XP","Bonus XP"\n',
+      ...gradingOverviews.map(
+        e =>
+          [
+            e.assessmentNumber,
+            e.assessmentName,
+            e.studentName,
+            e.studentUsername,
+            e.groupName,
+            e.submissionStatus,
+            e.gradingStatus,
+            e.questionCount,
+            e.gradedCount,
+            e.initialXp,
+            e.xpAdjustment,
+            e.currentXp,
+            e.maxXp,
+            e.xpBonus
+          ]
+            .map(field => `"${field}"`) // wrap each field in double quotes in case it contains a comma
+            .join(',') + '\n'
+      )
+    ],
+    { type: 'text/csv' }
+  );
+  const fileName = `SA submissions (${new Date().toISOString()}).csv`;
+
+  // code from https://github.com/ag-grid/ag-grid/blob/latest/grid-community-modules/csv-export/src/csvExport/downloader.ts
+  const element = document.createElement('a');
+  const url = win.URL.createObjectURL(content);
+  element.setAttribute('href', url);
+  element.setAttribute('download', fileName);
+  element.style.display = 'none';
+  document.body.appendChild(element);
+
+  element.dispatchEvent(
+    new MouseEvent('click', {
+      bubbles: false,
+      cancelable: true,
+      view: win
+    })
+  );
+
+  document.body.removeChild(element);
+
+  win.setTimeout(() => {
+    win.URL.revokeObjectURL(url);
+  }, 0);
+};
